backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
websockets==12.0
python-dateutil==2.8.2
pytz==2023.3
solders==0.19.4
base58==2.1.1
PyJWT==2.8.1
python-jose==3.3.0
cryptography==41.0.7
passlib==1.7.4
PyNaCl==1.5.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
